Add uprn-council code data and functionality to get a council code from a uprn
