+++ .gitignore
@@ -13,3 +13,4 @@ package-lock.json
 package.json
 pyvenv.cfg
 data/bdline_gpkg_gb
+data/onsud_apr_2025
